requirements.txt
aiohttp==2.2.5
arrow==0.12.1
async-timeout==3.0.0
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a0
feedparser==5.2.1
idna==2.6
multidict==4.3.1
psutil==5.4.5
pycparser==2.18
PyNaCl==1.1.2
python-dateutil==2.7.3
PyYAML==3.12
six==1.11.0
websockets==3.4
yarl==1.1.1
